+++ docs/aewx-metar-url.sh
@@ -3,7 +3,7 @@
 cd ~/Documents/Aerofly FS 2 # Replace with actual save data directory
 cp main.mcf main.bak
 
-METAR_URL=https://3960.org/metar/
-METAR_URL_END=.txt
+# Replace https://3960.org/metar/XXXX.txt with actual API key
+AEWX_URL=https://3960.org/metar/XXXX.txt
 read -p 'Enter ICAO string: ' ICAO
-aerofly-metar-url ${METAR_URL}${ICAO}${METAR_URL_END} --verbose
+aerofly-metar-fetch ${ICAO} ${AEWX_URL} --verbose